rework walk to proxy
