+++ src/api/java/com/minecolonies/api/util/EntityUtils.java
@@ -7,6 +7,7 @@ import net.minecraft.entity.EntityLiving;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.init.Blocks;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.Vec3i;
 import net.minecraft.world.World;
 import net.minecraftforge.common.util.FakePlayer;
 import org.jetbrains.annotations.NotNull;
@@ -27,6 +28,14 @@ public final class EntityUtils
      * How many blocks the citizen needs to stand safe.
      */
     private static final int    AIR_SPACE_ABOVE_TO_CHECK = 2;
+
+    /**
+     * Default range for moving to something until we stop.
+     */
+    private static final int    DEFAULT_MOVE_RANGE       = 3;
+    private static final int    TELEPORT_RANGE      = 512;
+    private static final double MIDDLE_BLOCK_OFFSET = 0.5D;
+    private static final int    SCAN_RADIUS         = 5;
 
     /**
      * Private constructor to hide the implicit public one.
@@ -214,5 +223,79 @@ public final class EntityUtils
     public static boolean tryMoveLivingToXYZ(@NotNull final EntityLiving living, final int x, final int y, final int z, final double speed)
     {
         return living.getNavigator().tryMoveToXYZ(x, y, z, speed);
+    }
+
+    /**
+     * {@link #isWorkerAtSiteWithMove(EntityCitizen, int, int, int, int)}.
+     *
+     * @param worker Worker to check
+     * @param x      X-coordinate
+     * @param y      Y-coordinate
+     * @param z      Z-coordinate
+     * @return True if worker is at site, otherwise false
+     */
+    public static boolean isLivingAtSiteWithMove(@NotNull final EntityLiving worker, final int x, final int y, final int z)
+    {
+        //Default range of 3 works better
+        //Range of 2 get some workers stuck
+        return isLivingAtSiteWithMove(worker, x, y, z, DEFAULT_MOVE_RANGE);
+    }
+
+    /**
+     * Checks if a worker is at his working site.
+     * If he isn't, sets it's path to the location.
+     *
+     * @param worker Worker to check
+     * @param x      X-coordinate
+     * @param y      Y-coordinate
+     * @param z      Z-coordinate
+     * @param range  Range to check in
+     * @return True if worker is at site, otherwise false.
+     */
+    public static boolean isLivingAtSiteWithMove(@NotNull final EntityLiving worker, final int x, final int y, final int z, final int range)
+    {
+        if (!isLivingAtSite(worker, x, y, z, TELEPORT_RANGE))
+        {
+            final BlockPos spawnPoint =
+                    Utils.scanForBlockNearPoint(worker.getEntityWorld(),
+                            new BlockPos(x, y, z),
+                            SCAN_RADIUS, SCAN_RADIUS, SCAN_RADIUS, 2,
+                            Blocks.AIR,
+                            Blocks.SNOW_LAYER,
+                            Blocks.TALLGRASS,
+                            Blocks.RED_FLOWER,
+                            Blocks.YELLOW_FLOWER,
+                            Blocks.CARPET);
+
+            worker.setLocationAndAngles(
+                    spawnPoint.getX() + MIDDLE_BLOCK_OFFSET,
+                    spawnPoint.getY(),
+                    spawnPoint.getZ() + MIDDLE_BLOCK_OFFSET,
+                    worker.rotationYaw,
+                    worker.rotationPitch);
+            return true;
+        }
+
+        if (!EntityUtils.isLivingAtSite(worker, x, y, z, range))
+        {
+            return false;
+        }
+        return true;
+    }
+
+    /**
+     * Returns whether or not the worker is within a specific range of his
+     * working site.
+     *
+     * @param worker Worker to check
+     * @param x      X-coordinate
+     * @param y      Y-coordinate
+     * @param z      Z-coordinate
+     * @param range  Range to check in
+     * @return True if worker is at site, otherwise false
+     */
+    public static boolean isLivingAtSite(@NotNull final EntityLiving worker, final int x, final int y, final int z, final int range)
+    {
+        return worker.getPosition().distanceSq(new Vec3i(x, y, z)) < MathUtils.square(range);
     }
 }
